+++ src/telemetry_data.h
@@ -345,7 +345,7 @@ class TelemetryData: public PacketData
          * 
          * \return 
          */
-        const char* PrintTime(uint8_t h, uint8_t m, uint8_t s);
+        const char* PrintTime(uint16_t h, uint8_t m, uint8_t s);
         /**
          * \brief 
          *
